FinalForm/requirements.txt
numpy
pandas
matplotlib
scikit-learn
scipy
joblib
h5py
astropy
Pillow
torch
timm
